[PE-1357] Add support for Network transitions (#51)
Currently, when we create/update/delete a connection, we immediately continue.
This may not be the case since these can be executed on a separate thread or handled
but a separate process.

Now when we create/update a network, we wait until it says that it's in the active state
before continuing. When we delete a connection, we make sure it is deleted before continuing.